requirements.txt
django
django-annoying
oauth2client
requests
